Use NO_GTYPE with many _WRAP_GERROR() and _WRAP_ENUMS() lines.
libgda-6.0 does not rnu glib-mkenums over all of the header
files anymore, so not all enums have *_get_type() functions.
This is apparently intentional though it was an unrelated
change in this commit:
https://git.gnome.org/browse/libgda/commit/libgda/Makefile.am?id=0111e1411cdaa9bca1169b367ef2ec26338c432b

It only prevents us from using these enums in a Glib::Value<>
so it is not a big problem.